Remove @browser and point out code blocks must be wrapped in ```.
When viewing the rendered markdown it's not clear that the code blocks need to be wrapped in ```. Make that explicit to avoid confusion for people who are just starting out and wondering why even the most basic "Hello World" isn't working.

Also, `bind @browser` is throwing an error "Invalid equality in update". Just `bind` seems to work.